Move various io-related source files to src/io/
