+++ src/editor/world.c
@@ -32,9 +32,9 @@
 #include "../world.h"
 #include "../legacy_board.h"
 #include "../idput.h"
-#include "../memfile.h"
 #include "../util.h"
-#include "../zip.h"
+#include "../io/memfile.h"
+#include "../io/zip.h"
 
 #include "../world_format.h"
