+++ src/io/fsafeopen.c
@@ -24,9 +24,9 @@
 #include <sys/stat.h>
 
 #include "fsafeopen.h"
-#include "const.h"
 #include "dir.h"
-#include "util.h"
+
+#include "../util.h"
 
 #ifndef __WIN32__
