+++ src/io/zip_reduce.h
@@ -32,7 +32,7 @@ __M_BEGIN_DECLS
 #include <inttypes.h>
 
 #include "bitstream.h"
-#include "dict.h"
+#include "zip_dict.h"
 #include "zip_stream.h"
 
 #include "../util.h"
