+++ src/world.c
@@ -42,17 +42,17 @@
 #include "error.h"
 #include "event.h"
 #include "extmem.h"
-#include "fsafeopen.h"
 #include "game_player.h"
 #include "graphics.h"
 #include "idput.h"
-#include "memfile.h"
 #include "robot.h"
 #include "sprite.h"
 #include "str.h"
 #include "util.h"
 #include "window.h"
-#include "zip.h"
+#include "io/memfile.h"
+#include "io/fsafeopen.h"
+#include "io/zip.h"
 
 #include "audio/audio.h"
 #include "audio/sfx.h"
